third_party/WebKit/LayoutTests/http/tests/devtools/console/console-log-eval-syntax-error-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that syntax errors in eval are logged into console, contains correct link and doesn't cause browser crash.

VM:1 Uncaught SyntaxError: Unexpected token }
    at foo (<anonymous>:13:11)
foo @ VM:13
setTimeout (async)
(anonymous) @ console-log-eval-syntax-error.js:21
boo.js:2 Uncaught SyntaxError: Unexpected token }
    at boo (<anonymous>:17:11)
boo @ VM:17
setTimeout (async)
(anonymous) @ console-log-eval-syntax-error.js:24
